requirements.txt
numpy==2.3.1
pandas==2.3.0
polars==1.31.0
py4j==0.10.9.9
pyarrow==20.0.0
pyspark==4.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
tzdata==2025.2
